+++ src/IDBFactory.js
@@ -59,13 +59,14 @@ IDBFactory.prototype.open = function (name, version) {
     if (arguments.length === 0) {
         throw new TypeError('Database name is required');
     }
-    if (arguments.length >= 2) {
+    if (version !== undefined) {
         version = Number(version);
         if (isNaN(version) || !isFinite(version) ||
             version >= 0x20000000000000 || // 2 ** 53
             version < 1) { // The spec only mentions version==0 as throwing, but W3C tests fail with these
             throw new TypeError('Invalid database version: ' + version);
         }
+        version = Math.floor(version);
     }
     name = String(name); // cast to a string
     const sqlSafeName = util.escapeSQLiteStatement(name);